# Task2/e-transport_back/swagger-spec.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/rental/{id}": {
            "put": {
                "operationId": "RentalController_updateRental",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRentalDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rental updated successfully."
                    },
                    "404": {
                        "description": "Rental not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "rental"
                ]
            },
            "delete": {
                "operationId": "RentalController_deleteRental",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rental deleted successfully."
                    },
                    "404": {
                        "description": "Rental not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "rental"
                ]
            }
        },
        "/payment": {
            "get": {
                "operationId": "PaymentController_findAllPayment",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "List of all payments returned successfully."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "payment"
                ]
            },
            "post": {
                "operationId": "PaymentController_createPayment",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payment created successfully."
                    },
                    "400": {
                        "description": "Invalid input data."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "payment"
                ]
            }
        },
        "/payment/{id}": {
            "put": {
                "operationId": "PaymentController_updatePayment",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePaymentDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment updated successfully."
                    },
                    "404": {
                        "description": "Payment not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "payment"
                ]
            },
            "delete": {
                "operationId": "PaymentController_deletePayment",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment deleted successfully."
                    },
                    "404": {
                        "description": "Payment not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "payment"
                ]
            }
        },
...
        "/vehicle/{id}": {
            "put": {
                "operationId": "VehicleController_updateVehicle",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateVehicleDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vehicle updated successfully."
                    },
                    "404": {
                        "description": "Vehicle not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "vehicle"
                ]
            },
            "delete": {
                "operationId": "VehicleController_deleteVehicle",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicle deleted successfully."
                    },
                    "404": {
                        "description": "Vehicle not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "vehicle"
                ]
            }
        },
...
        "/rental-vehicle": {
            "get": {
                "operationId": "RentalVehicleController_findAllRentalVehicle",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "List of all rental vehicles returned successfully."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "rental-vehicle"
                ]
            },
            "post": {
                "operationId": "RentalVehicleController_createRentalVehicle",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RentalVehicleDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rental vehicle created successfully."
                    },
                    "400": {
                        "description": "Invalid input data."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "rental-vehicle"
                ]
            }
        },
        "/rental-vehicle/{id}": {
            "put": {
                "operationId": "RentalVehicleController_updateRentalVehicle",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRentalVehicleDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rental vehicle updated successfully."
                    },
                    "404": {
                        "description": "Rental vehicle not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "rental-vehicle"
                ]
            },
            "delete": {
                "operationId": "RentalVehicleController_deleteRentalVehicle",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rental vehicle deleted successfully."
                    },
                    "404": {
                        "description": "Rental vehicle not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "rental-vehicle"
                ]
            }
        },
...
    },
    "info": {
        "title": "e-transport",
        "description": "The e-transport API description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
...
            "UpdateRentalDto": {
                "type": "object",
                "properties": {
                    "isActive": {
                        "type": "boolean"
                    },
                    "dateRented": {
                        "type": "string"
                    },
                    "dateReturned": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "string"
                    },
                    "distance": {
                        "type": "number"
                    },
                    "avgSpeed": {
                        "type": "number"
                    },
                    "maxSpeed": {
                        "type": "number"
                    },
                    "energyConsumed": {
                        "type": "number"
                    }
                }
            },
            "PaymentDto": {
                "type": "object",
                "properties": {
                    "paymentMethod": {
                        "type": "string",
                        "enum": [
                            "VISA",
                            "PAYPAL",
                            "MASTERCARD",
                            "OTHER"
                        ]
                    },
                    "amount": {
                        "type": "string"
                    },
                    "date": {
                        "type": "string"
                    },
                    "rentalId": {
                        "type": "string"
                    }
                },
                "required": [
                    "paymentMethod",
                    "amount",
                    "date",
                    "rentalId"
                ]
            },
            "UpdatePaymentDto": {
                "type": "object",
                "properties": {
                    "paymentMethod": {
                        "type": "string",
                        "enum": [
                            "VISA",
                            "PAYPAL",
                            "MASTERCARD",
                            "OTHER"
                        ]
                    },
                    "name": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "string"
                    },
                    "date": {
                        "type": "string"
                    },
                    "rentalId": {
                        "type": "string"
                    }
                },
                "required": [
                    "paymentMethod"
                ]
            },
...
            "UpdateVehicleDto": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "FREE",
                            "INUSE",
                            "NOTAVAILABLE",
                            "BROKEN",
                            "REPAIR"
                        ]
                    },
                    "runnedDistance": {
                        "type": "number"
                    },
                    "releaseDate": {
                        "type": "string"
                    },
                    "currentLocation": {
                        "type": "string"
                    }
                },
                "required": [
                    "status"
                ]
            },
...
            "RentalVehicleDto": {
                "type": "object",
                "properties": {
                    "vehicleId": {
                        "type": "string"
                    },
                    "rentalId": {
                        "type": "string"
                    }
                },
                "required": [
                    "vehicleId",
                    "rentalId"
                ]
            },
            "UpdateRentalVehicleDto": {
                "type": "object",
                "properties": {
                    "vehicleId": {
                        "type": "string"
                    },
                    "rentalId": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}